Added pydoc for gui
