+++ documentation/sphinx_doc/gui.rst
@@ -1,0 +1,5 @@
+GUI
+===
+
+.. automodule:: hal_gui.gui
+   :members:
