+++ documentation/sphinx_doc/hal_gui.rst
@@ -1,0 +1,7 @@
+hal_gui Documentation
+=====================
+
+.. toctree::
+   :maxdepth: 1
+
+   gui
